refactor(archtest): migrate SVCTOKEN-CALLER-CELL to scanner framework
283 → 163 lines. Replace manual filepath.WalkDir + ReadFile +
ParseFile fail-loud template with DirsScope(IncludeTests) + EachFile
+ []Diagnostic + Report. Move findAllGoFilesInDir to helpers_test.go
(still needed by no_deleted_auth_symbols and
no_test_service_context_in_production tests).

Refs: PR408-FU-SCANNER-SHARED-FRAMEWORK-01
