+++ tools/archtest/svctoken_caller_cell_test.go
@@ -21,16 +21,11 @@ import (
 	"go/ast"
 	"go/parser"
 	"go/token"
-	"os"
-	"path/filepath"
 	"regexp"
-	"sort"
-	"strings"
 	"testing"
 
-	"github.com/stretchr/testify/require"
-
 	"github.com/ghbvf/gocell/kernel/metadata"
+	"github.com/ghbvf/gocell/tools/archtest/internal/scanner"
 )
 
 // ruleSvctokenCallerCellRequired01 is the archtest rule identifier; not a credential.
@@ -59,45 +54,84 @@ func TestSVCTOKEN_CALLER_CELL_REQUIRED_01(t *testing.T) {
 	// and must declare a known callerCell. Direct walk is correct here:
 	// the rule scope is wider than cell-managed code (e.g.,
 	// examples/ssobff/walkthrough_test.go is non-cell example code).
-	searchDirs := []string{
-		filepath.Join(root, "runtime"),
-		filepath.Join(root, "cells"),
-		filepath.Join(root, "cmd"),
-		filepath.Join(root, "examples"),
-		filepath.Join(root, "tests"),
-	}
-	var allFiles []string
-	for _, dir := range searchDirs {
-		ff, err := findAllGoFilesInDir(dir)
-		if os.IsNotExist(err) {
-			continue
-		}
-		if err != nil {
-			t.Fatalf("walking %s: %v", dir, err)
-		}
-		allFiles = append(allFiles, ff...)
-	}
+	scope := scanner.DirsScope(root,
+		[]string{"runtime", "cells", "cmd", "examples", "tests"},
+		scanner.IncludeTests(),
+	)
 
-	var violations []string
-	for _, f := range allFiles {
-		rel, _ := filepath.Rel(root, f)
-		rel = filepath.ToSlash(rel)
-		hits, scanErr := scanGenerateServiceTokenCallSites(f, rel, knownCells)
-		require.NoError(t, scanErr)
-		violations = append(violations, hits...)
-	}
+	var diags []scanner.Diagnostic
+	scanner.EachFile(t, scope, parser.SkipObjectResolution, func(t *testing.T, fc scanner.FileContext) {
+		authAliases := authPackageAliases(fc.File)
+		if len(authAliases) == 0 {
+			return // file does not import runtime/auth
+		}
+		ast.Inspect(fc.File, func(n ast.Node) bool {
+			call, ok := n.(*ast.CallExpr)
+			if !ok {
+				return true
+			}
+			if !isAuthGenerateServiceTokenCall(call, authAliases) {
+				return true
+			}
+			pos := fc.Fset.Position(call.Pos())
 
-	sort.Strings(violations)
-	for _, v := range violations {
-		t.Log(v)
-	}
-	if len(violations) > 0 {
-		t.Errorf("%s: %d auth.GenerateServiceToken call sites with invalid/missing callerCell.\n"+
-			"The second argument must be a non-empty string literal matching ^[a-z][a-z0-9-]*$\n"+
-			"and must be a known cell ID from cells/ or actors.yaml.\n"+
-			"Known cells: %v",
-			ruleSvctokenCallerCellRequired01, len(violations), sortedKeys(knownCells))
-	}
+			// The 4-part signature is GenerateServiceToken(ring, callerCell, method, path, query, ts).
+			// callerCell is argument index 1 (0-based).
+			if len(call.Args) < 2 {
+				diags = append(diags, scanner.Diagnostic{
+					Rel:     fc.Rel,
+					Line:    pos.Line,
+					Message: "auth.GenerateServiceToken called with fewer than 2 arguments — missing callerCell",
+				})
+				return true
+			}
+
+			arg1 := call.Args[1]
+			lit, ok := arg1.(*ast.BasicLit)
+			if !ok || lit.Kind != token.STRING {
+				diags = append(diags, scanner.Diagnostic{
+					Rel:     fc.Rel,
+					Line:    pos.Line,
+					Message: "auth.GenerateServiceToken second argument (callerCell) must be a string literal",
+				})
+				return true
+			}
+
+			// Strip quotes.
+			callerCell := lit.Value
+			if len(callerCell) >= 2 && callerCell[0] == '"' && callerCell[len(callerCell)-1] == '"' {
+				callerCell = callerCell[1 : len(callerCell)-1]
+			}
+
+			if callerCell == "" {
+				diags = append(diags, scanner.Diagnostic{
+					Rel:     fc.Rel,
+					Line:    pos.Line,
+					Message: "auth.GenerateServiceToken callerCell must not be empty",
+				})
+				return true
+			}
+
+			if !cellIDRegex.MatchString(callerCell) {
+				diags = append(diags, scanner.Diagnostic{
+					Rel:     fc.Rel,
+					Line:    pos.Line,
+					Message: fmt.Sprintf("auth.GenerateServiceToken callerCell %q does not match ^[a-z][a-z0-9-]*$", callerCell),
+				})
+				return true
+			}
+
+			if !knownCells[callerCell] {
+				diags = append(diags, scanner.Diagnostic{
+					Rel:     fc.Rel,
+					Line:    pos.Line,
+					Message: fmt.Sprintf("auth.GenerateServiceToken callerCell %q is not a known cell ID — register it in cells/ or actors.yaml", callerCell),
+				})
+			}
+			return true
+		})
+	})
+	scanner.Report(t, ruleSvctokenCallerCellRequired01, diags)
 }
 
 // discoverKnownCells returns the set of valid caller cell IDs from
@@ -122,85 +156,6 @@ func discoverKnownCells(t *testing.T, root string) map[string]bool {
 		}
 	}
 	return known
-}
-
-// scanGenerateServiceTokenCallSites parses a .go file and returns violation
-// strings for any auth.GenerateServiceToken call where the second argument
-// (callerCell) is not a valid known cell-ID string literal.
-func scanGenerateServiceTokenCallSites(path, rel string, knownCells map[string]bool) ([]string, error) {
-	data, err := os.ReadFile(filepath.Clean(path))
-	if err != nil {
-		return nil, err
-	}
-	fset := token.NewFileSet()
-	f, err := parser.ParseFile(fset, path, data, parser.SkipObjectResolution)
-	if err != nil {
-		return nil, fmt.Errorf("parse %s: %w", path, err)
-	}
-
-	authAliases := authPackageAliases(f)
-	if len(authAliases) == 0 {
-		return nil, nil // file does not import runtime/auth
-	}
-
-	var violations []string
-	ast.Inspect(f, func(n ast.Node) bool {
-		call, ok := n.(*ast.CallExpr)
-		if !ok {
-			return true
-		}
-		if !isAuthGenerateServiceTokenCall(call, authAliases) {
-			return true
-		}
-		pos := fset.Position(call.Pos())
-
-		// The 4-part signature is GenerateServiceToken(ring, callerCell, method, path, query, ts).
-		// callerCell is argument index 1 (0-based).
-		if len(call.Args) < 2 {
-			violations = append(violations, fmt.Sprintf(
-				"%s:%d: auth.GenerateServiceToken called with fewer than 2 arguments — missing callerCell",
-				rel, pos.Line))
-			return true
-		}
-
-		arg1 := call.Args[1]
-		lit, ok := arg1.(*ast.BasicLit)
-		if !ok || lit.Kind != token.STRING {
-			violations = append(violations, fmt.Sprintf(
-				"%s:%d: auth.GenerateServiceToken second argument (callerCell) must be a string literal",
-				rel, pos.Line))
-			return true
-		}
-
-		// Strip quotes.
-		callerCell := lit.Value
-		if len(callerCell) >= 2 && callerCell[0] == '"' && callerCell[len(callerCell)-1] == '"' {
-			callerCell = callerCell[1 : len(callerCell)-1]
-		}
-
-		if callerCell == "" {
-			violations = append(violations, fmt.Sprintf(
-				"%s:%d: auth.GenerateServiceToken callerCell must not be empty",
-				rel, pos.Line))
-			return true
-		}
-
-		if !cellIDRegex.MatchString(callerCell) {
-			violations = append(violations, fmt.Sprintf(
-				"%s:%d: auth.GenerateServiceToken callerCell %q does not match ^[a-z][a-z0-9-]*$",
-				rel, pos.Line, callerCell))
-			return true
-		}
-
-		if !knownCells[callerCell] {
-			violations = append(violations, fmt.Sprintf(
-				"%s:%d: auth.GenerateServiceToken callerCell %q is not a known cell ID — "+
-					"register it in cells/ or actors.yaml",
-				rel, pos.Line, callerCell))
-		}
-		return true
-	})
-	return violations, nil
 }
 
 // isAuthGenerateServiceTokenCall reports whether call is a call expression
@@ -248,35 +203,3 @@ func authPackageAliases(f *ast.File) map[string]struct{} {
 	return out
 }
 
-// findAllGoFilesInDir walks dir and returns all .go files (including _test.go).
-func findAllGoFilesInDir(dir string) ([]string, error) {
-	var files []string
-	err := filepath.WalkDir(dir, func(path string, d os.DirEntry, err error) error {
-		if err != nil {
-			return err
-		}
-		if d.IsDir() {
-			switch d.Name() {
-			case "vendor", ".git", "generated", "testdata":
-				return filepath.SkipDir
-			}
-			return nil
-		}
-		if strings.HasSuffix(path, ".go") {
-			files = append(files, path)
-		}
-		return nil
-	})
-	sort.Strings(files)
-	return files, err
-}
-
-// sortedKeys returns a sorted slice of map keys for diagnostic output.
-func sortedKeys(m map[string]bool) []string {
-	keys := make([]string, 0, len(m))
-	for k := range m {
-		keys = append(keys, k)
-	}
-	sort.Strings(keys)
-	return keys
-}
